requirements.txt
flask==2.0.1
openai==1.55.0
python-dotenv==1.0.1 